Feat: Add initial project structure and settings
Spring Boot (Java 21)
- 6 core dependencies (Web, JPA, Security, Validation, etc.)
- 7 core packages (config, entity, controller, etc.)
- MySQL DB connection in application.yml
- Standard .gitignore file

+++ src/main/java/teamprojects/demo/repository/PointHistoryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import teamprojects.demo.entity.PointHistory;
 import teamprojects.demo.entity.User;
 
-public interface PointHistoryRepository extends JpaRepository<PointHistory, Long> {
+public interface PointHistoryRepository extends JpaRepository<PointHistory, Integer> {
 
     // (API 4-4: '내 포인트 내역 조회' 페이징 처리 시 사용)
     Page<PointHistory> findByUserOrderByCreatedAtDesc(User user, Pageable pageable);
