+++ src/main/java/teamprojects/demo/repository/PraiseRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import teamprojects.demo.entity.Praise;
 import teamprojects.demo.entity.User;
 
-public interface PraiseRepository extends JpaRepository<Praise, Long> {
+public interface PraiseRepository extends JpaRepository<Praise, Integer> {
 
     // (API 4-1: 마이페이지 칭찬 횟수 요약 시 사용)
     long countByReceiver(User receiver);
