+++ src/main/java/teamprojects/demo/repository/ReliabilityHistoryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import teamprojects.demo.entity.ReliabilityHistory;
 import teamprojects.demo.entity.User;
 
-public interface ReliabilityHistoryRepository extends JpaRepository<ReliabilityHistory, Long> {
+public interface ReliabilityHistoryRepository extends JpaRepository<ReliabilityHistory, Integer> {
 
     // (API 4-5: '내 신뢰도 내역 조회' 페이징 처리 시 사용)
     Page<ReliabilityHistory> findByUserOrderByCreatedAtDesc(User user, Pageable pageable);
