+++ src/main/java/teamprojects/demo/repository/StudyCategoryRepository.java
@@ -3,6 +3,6 @@ package teamprojects.demo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import teamprojects.demo.entity.StudyCategory;
 
-public interface StudyCategoryRepository extends JpaRepository<StudyCategory, Long> {
+public interface StudyCategoryRepository extends JpaRepository<StudyCategory, Integer> {
 
 }
