+++ src/main/java/teamprojects/demo/repository/TodoListRepository.java
@@ -8,7 +8,7 @@ import teamprojects.demo.entity.User;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface TodoListRepository extends JpaRepository<TodoList, Long> {
+public interface TodoListRepository extends JpaRepository<TodoList, Integer> {
 
     // (API 5-2: 특정 날짜의 TODO 목록 조회 시 사용)
     // Service단에서 targetDate의 시작(00:00:00)과 끝(23:59:59) 시간을
