+++ src/main/java/teamprojects/demo/repository/UserProfileRepository.java
@@ -3,7 +3,7 @@ package teamprojects.demo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import teamprojects.demo.entity.UserProfile;
 
-public interface UserProfileRepository extends JpaRepository<UserProfile, Long> {
+public interface UserProfileRepository extends JpaRepository<UserProfile, Integer> {
 
     boolean existsByUsername(String username);
 }
